Separated oracles and type_checker module + added theory oracle in env
